refactor: create reducer.ts file
